+++ extractor/extractor.go
@@ -1,4 +1,4 @@
-package main
+package extractor
 
 import (
 	"bufio"
