+++ extractor/helpers.go
@@ -1,4 +1,4 @@
-package main
+package extractor
 
 import (
 	"encoding/json"
